+++ Grbl_Esp32/src/MotionControl.cpp
@@ -166,14 +166,14 @@ void mc_arc(float*            target,
     }
 
     // CCW angle between position and target from circle center. Only one atan2() trig computation required.
-    float angular_travel = atan2(r_axis0 * rt_axis1 - r_axis1 * rt_axis0, r_axis0 * rt_axis0 + r_axis1 * rt_axis1);
+    float angular_travel = float(atan2(r_axis0 * rt_axis1 - r_axis1 * rt_axis0, r_axis0 * rt_axis0 + r_axis1 * rt_axis1));
     if (is_clockwise_arc) {  // Correct atan2 output per direction
         if (angular_travel >= -ARC_ANGULAR_TRAVEL_EPSILON) {
-            angular_travel -= 2 * M_PI;
+            angular_travel -= 2 * float(M_PI);
         }
     } else {
         if (angular_travel <= ARC_ANGULAR_TRAVEL_EPSILON) {
-            angular_travel += 2 * M_PI;
+            angular_travel += 2 * float(M_PI);
         }
     }
 
@@ -183,7 +183,8 @@ void mc_arc(float*            target,
     // (2x) arc_tolerance. For 99% of users, this is just fine. If a different arc segment fit
     // is desired, i.e. least-squares, midpoint on arc, just change the mm_per_arc_segment calculation.
     // For the intended uses of Grbl, this value shouldn't exceed 2000 for the strictest of cases.
-    uint16_t segments = floor(fabs(0.5 * angular_travel * radius) / sqrt(mconfig->_arcTolerance * (2 * radius - mconfig->_arcTolerance)));
+    uint16_t segments =
+        uint16_t(floor(fabs(0.5 * angular_travel * radius) / sqrt(mconfig->_arcTolerance * (2 * radius - mconfig->_arcTolerance))));
     if (segments) {
         // Multiply inverse feed_rate to compensate for the fact that this movement is approximated
         // by a number of discrete segments. The inverse feed_rate should be correct for the sum of
@@ -220,8 +221,8 @@ void mc_arc(float*            target,
            This is important when there are successive arc motions.
         */
         // Computes: cos_T = 1 - theta_per_segment^2/2, sin_T = theta_per_segment - theta_per_segment^3/6) in ~52usec
-        float cos_T = 2.0 - theta_per_segment * theta_per_segment;
-        float sin_T = theta_per_segment * 0.16666667 * (cos_T + 4.0);
+        float cos_T = 2.0f - theta_per_segment * theta_per_segment;
+        float sin_T = theta_per_segment * 0.16666667f * (cos_T + 4.0f);
         cos_T *= 0.5;
         float    sin_Ti;
         float    cos_Ti;
@@ -239,8 +240,8 @@ void mc_arc(float*            target,
             } else {
                 // Arc correction to radius vector. Computed only every N_ARC_CORRECTION increments. ~375 usec
                 // Compute exact location by applying transformation matrix from initial radius vector(=-offset).
-                cos_Ti  = cos(i * theta_per_segment);
-                sin_Ti  = sin(i * theta_per_segment);
+                cos_Ti  = float(cos(i * theta_per_segment));
+                sin_Ti  = float(sin(i * theta_per_segment));
                 r_axis0 = -offset[axis_0] * cos_Ti + offset[axis_1] * sin_Ti;
                 r_axis1 = -offset[axis_0] * sin_Ti - offset[axis_1] * cos_Ti;
                 count   = 0;
