+++ Grbl_Esp32/src/Protocol.cpp
@@ -222,7 +222,7 @@ void protocol_main_loop() {
         }
         // check to see if we should disable the stepper drivers ... esp32 work around for disable in main loop.
         if (stepper_idle && config->_idleTime != 255) {
-            if (esp_timer_get_time() > stepper_idle_counter) {
+            if (uint64_t(esp_timer_get_time()) > stepper_idle_counter) {
                 config->_axes->set_disable(true);
             }
         }
